pyproject.toml
[tool.poetry]
name = "quizbot"
version = "0.1.0"
description = ""
authors = ["balancy <balancy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.8.1"
environs = "^9.3.4"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
